feat: componentise the language selector and add it to the login page
